Critical Fix for Timer & Enhancements
Turns out the timer was not working due to some bug with the setTimeout
function. I rewrote it slightly different which should fix the problem.
Also added a 'refreshed' time to display the last time the widget
updated the data. This is different from the 'updated' date which
reflects the data freshness and is provided by coinmarketcap via their
API. Lastly, the date formats were shortened because only the time is
really relevant.